blog-details edited

+++ src/main/java/com/book/store/model/Book.java
@@ -21,7 +21,7 @@ public class Book {
     private String secondImagePath;
     private String language;
     private LocalDate writeDate;
-    private double avarageRating;
+    private int avarageRating;
     private Author author;
     private List<Category> categoryList;
     private List<Review> reviewList;
